server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-auth==2.27.0
gunicorn==21.2.0
